Server/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
bidict==0.21.2
certifi==2020.12.5
chardet==4.0.0
idna==2.10
loguru==0.5.3
multidict==5.1.0
python-engineio==4.0.1
python-socketio==5.1.0
requests==2.25.1
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.4
websocket-client==0.58.0
yarl==1.6.3
